requirements.txt
keras_vggface==0.6
scikit-learn==1.5.1
streamlit==1.36.0
Pillow==10.4.0
opencv-python-headless==4.10.0
mtcnn==0.1.0
numpy==1.26.4
